Task 15: Add observability, metrics and security
- Add structured logging using tracing and tracing-subscriber
  - Daily rotating log files in LOG_DIR (default: logs/)
  - JSON format with target, thread IDs, file and line numbers
  - Panic handler that logs panic location and message
- Add request tracing with tower-http TraceLayer
- Add metrics system with counters for HTTP, auth, and business events
  - GET /metrics endpoint returns JSON metrics
- Add health check endpoint at GET /health
- Add audit logging system with in-memory storage
- Add request ID generation using UUID v4
- Add AppState with Metrics and AuditLogger to server
- Fix clippy issues in auth and cache crates
- Format code with cargo fmt

+++ backend/crates/common/src/cache/valkey.rs
@@ -59,7 +59,7 @@ impl ValkeyCache {
 impl Cache for ValkeyCache {
     async fn set(&self, key: &str, value: &str, ttl: Duration) -> Result<(), CacheError> {
         let mut conn = self.get_conn().await?;
-        let ttl_secs = ttl.as_secs() as u64;
+        let ttl_secs = ttl.as_secs();
 
         conn.set_ex::<_, _, ()>(key, value, ttl_secs)
             .await
